Refactor: Range 도메인을 Board에서 Community로 이동 및 Notice range 필드 수정
