--- src/main/java/com/knud4/an/board/Range.java
@@ -1,5 +1,0 @@
-package com.knud4.an.board;
-
-public enum Range {
-    ALL, LINE
-}
