+++ src/main/java/com/knud4/an/community/entity/Range.java
@@ -1,0 +1,5 @@
+package com.knud4.an.community.entity;
+
+public enum Range {
+    ALL, LINE
+}
